+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Versioning].
 
 ## [Unreleased] <!-- #release:date -->
 
+* Support building on Windows (#16).
+
 ## [0.3.1+1.19.2] - 2021-12-22
 
 * Incorporate a patch to remove an errant linker flag that was breaking
@@ -78,3 +80,5 @@ Initial release.
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 [krb5/krb5#1233]: https://github.com/krb5/krb5/pull/1233
+
+[#16]: https://github.com/MaterializeInc/rust-krb5-src/issues/16